refactor: extract Stars component into shared Stars.jsx
- Create client/src/components/Stars.jsx with useId() for stable clip path IDs
- Remove inline Stars definition from HomePage.jsx
- Remove inline Stars definition from ProductPage.jsx
- Both pages now import from shared component

Fixes #240
